Struct sgx_tdh::SgxDhMsg3

pub struct SgxDhMsg3 {
    pub cmac: [u8; 16],
    pub msg3_body: SgxDhMsg3Body,
}

Type for MSG3 used in DH secure session establishment.

Fields
cmac: [u8; 16] msg3_body: SgxDhMsg3Body
Methods

impl SgxDhMsg3

fn new() -> Self

Create a SgxDhMsg3 with default values.

fn calc_raw_sealed_data_size(&self) -> u32

Calculate the size of sgx_dh_msg3_t converted from SgxDhMsg3, really add the size of struct sgx_dh_msg3_t and msg3_body.additional_prop.

Return value

The size of sgx_dh_msg3_t needed.

unsafe fn to_raw_dh_msg3_t(
&self,
p: *mut sgx_dh_msg3_t,
len: u32
) -> Option<*mut sgx_dh_msg3_t>

Convert SgxDhMsg3 to sgx_dh_msg3_t, this is an unsafe function.

Parameters

p

The pointer of a sgx_dh_msg3_t buffer to save the buffer of SgxDhMsg3.

len

The size of the sgx_dh_msg3_t buffer.

Return value

Some(*mut sgx_dh_msg3_t)

Indicates the conversion is successfully. The return value is the mutable pointer of sgx_dh_msg3_t.

None

The parameters p and len are not available for the conversion.

unsafe fn from_raw_dh_msg3_t(p: *mut sgx_dh_msg3_t, len: u32) -> Option<Self>

Convert sgx_dh_msg3_t to SgxDhMsg3, this is an unsafe function.

Parameters

p

The pointer of a sgx_dh_msg3_t buffer.

len

The size of the sgx_dh_msg3_t buffer.

Return value

Some(SgxDhMsg3)

Indicates the conversion is successfully. The return value is SgxDhMsg3.

None

The parameters p and len are not available for the conversion.